Create README.md
Background: 
The client, a player in the media space, was eager to understand how people were interacting with each other on Twitter. He had collected a dataset of over 8 million records and wanted to derive meaningful insights into the behavior of Twitter users. He divided his requirements into four key steps: exploratory data analysis, creation of communities based on user interactions, sentiment analysis of users within each community, and identifying unusual behavior that may indicate bot accounts.

Problem: 
The size of the data posed a significant challenge, as existing algorithms for community detection were computationally expensive. Additionally, the data had many inconsistencies and missing values, with tweets in multiple languages (the task was restricted to analyzing tweets in German).

Solution: 
We began by carefully analyzing the data structure, identifying missing values, and exploring patterns such as the number of tweets, retweets, and replies, location of tweets, active users, most frequent tweets, and hashtags and mentioned accounts. This gave us a deep understanding of the data, which was essential for successfully applying machine learning models. Next, we used state-of-the-art algorithms like Leiden and Louvain to identify communities based on how users interacted with each other. We optimized the modularity score by experimenting with various parameters. Once the client was satisfied with the community detection, we moved on to sentiment analysis to see which communities were tweeting more positively or negatively. By combining sentiment analysis with hashtags used within the communities, we were able to identify hashtags with positive and negative connotations. Finally, we tackled the task of identifying unusual behavior that may indicate bot accounts. Following the client's guidelines, we looked for accounts that were created recently and had high numbers of retweets but few original tweets. We were able to successfully execute this step, much to the delight of our client.
Outcome: The client was thrilled with the results and impressed by the effort we put into executing the project. He was satisfied with the insightful analysis of user behavior on Twitter and appreciated the thorough examination of the data.